:fire: Long -> Integer로 수정

+++ src/main/java/com/vincent/domain/building/controller/dto/BuildingResponseDto.java
@@ -14,7 +14,7 @@ public class BuildingResponseDto {
     @AllArgsConstructor
     public static class BuildingInfo {
 
-        Long buildingId;
+        Integer buildingId;
         String buildingName;
         String buildingImage;
         String buildingAddress;
